Split IdentityRepository.Wrapper class out of IdentityRepository interface since it is only intended for internal use.
